memset_s: check value>255, use int
harmonize the API with C11.
use the libc variant only with
defined HAVE_MEMSET_S && defined HAVE_C11 && defined WANT_C11

Closes #27

+++ include/safe_mem_lib.h
@@ -81,7 +81,7 @@ memmove_s(void *dest, rsize_t  dmax,
 /* set bytes. now __STDC_WANT_LIB_EXT1__ >= 1 compatible */
 #if !(defined(__STDC_WANT_LIB_EXT1__) && (__STDC_WANT_LIB_EXT1__ >= 1))
 EXTERN errno_t
-memset_s(void *dest, rsize_t dmax, uint8_t value, rsize_t n);
+memset_s(void *dest, rsize_t dmax, int value, rsize_t n);
 #endif
 
 #ifndef SAFECLIB_DISABLE_EXTENSIONS
